# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Packwerk B2B Packaging Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [flexible, rigid, boxes, ecommerce, rolls, accessories, sustainable, premium]
        - name: is_eco
          in: query
          schema:
            type: boolean
        - name: is_smartstock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                  total:
                    type: integer
                required: [data, total]
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
...
  /quotes:
    post:
      operationId: submitQuote
      tags: [quotes]
      summary: Submit a new quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  quote_id:
                    type: string
                  id:
                    type: string
                required: [quote_id, id]
        "400":
          description: Invalid request
  /quotes/{quoteId}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote by quote_id
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
...
  /design-requests:
    post:
      operationId: createDesignRequest
      tags: [design]
      summary: Create a design request after payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDesignRequest"
      responses:
        "201":
          description: Design request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  design_id:
                    type: string
                  id:
                    type: string
                required: [design_id, id]
  /sample-requests:
    post:
      operationId: createSampleRequest
      tags: [samples]
      summary: Create a sample request after payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSampleRequest"
      responses:
        "201":
          description: Sample request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sample_id:
                    type: string
                  id:
                    type: string
                required: [sample_id, id]
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [flexible, rigid, boxes, ecommerce, rolls, accessories, sustainable, premium]
        description:
          type: string
        use_case:
          type: string
        price_min:
          type: number
        price_max:
          type: number
        moq:
          type: integer
        moq_unit:
          type: string
        is_smartstock:
          type: boolean
        is_eco:
          type: boolean
        sample_tier:
          type: string
          enum: [standard, premium, complex, not_required]
        sample_price:
          type: integer
        image_url:
          type: string
          nullable: true
        specs:
          type: object
        compliance_certs:
          type: array
          items:
            type: string
        delivery_days_india:
          type: integer
        delivery_days_global:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
      required: [id, name, slug, category, description, use_case, price_min, price_max, moq, moq_unit, is_smartstock, is_eco, sample_tier, sample_price, delivery_days_india, delivery_days_global, is_active]
    CreateQuoteRequest:
      type: object
      properties:
        contact_name:
          type: string
        company_name:
          type: string
        email:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: string
              product_name:
                type: string
              quantity:
                type: integer
              variants:
                type: string
              artwork_status:
                type: string
              sample_requested:
                type: boolean
              sample_tier:
                type: string
            required: [product_id, product_name, quantity]
        delivery_country:
          type: string
        delivery_pincode:
          type: string
        preferred_timeline:
          type: string
          enum: [urgent, standard, flexible]
        notes:
          type: string
        total_estimated_min:
          type: number
        total_estimated_max:
          type: number
      required: [contact_name, company_name, email, phone, items, delivery_country, preferred_timeline]
    QuoteRequest:
      type: object
      properties:
        id:
          type: string
        quote_id:
          type: string
        contact_name:
          type: string
        company_name:
          type: string
        email:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            type: object
        delivery_country:
          type: string
        delivery_pincode:
          type: string
          nullable: true
        preferred_timeline:
          type: string
        notes:
          type: string
          nullable: true
        total_estimated_min:
          type: number
          nullable: true
        total_estimated_max:
          type: number
          nullable: true
        status:
          type: string
          enum: [submitted, reviewing, quoted, accepted, rejected]
        rejection_reason:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, quote_id, contact_name, company_name, email, phone, items, delivery_country, preferred_timeline, status]
...
    CreateDesignRequest:
      type: object
      properties:
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        product_type:
          type: string
        product_id:
          type: string
        brand_colors:
          type: string
        logo_url:
          type: string
        brand_description:
          type: string
        notes:
          type: string
        is_rush:
          type: boolean
        amount_paid:
          type: integer
        razorpay_payment_id:
          type: string
      required: [contact_name, email, phone, product_type, amount_paid]
...
    CreateSampleRequest:
      type: object
      properties:
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        product_id:
          type: string
        sample_tier:
          type: string
          enum: [standard, premium, complex]
        amount_paid:
          type: integer
        razorpay_payment_id:
          type: string
      required: [contact_name, email, phone, product_id, sample_tier, amount_paid]
...