# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Packwerk B2B Packaging Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from "./productCategory";
import type { ProductSampleTier } from "./productSampleTier";
import type { ProductSpecs } from "./productSpecs";

export interface Product {
  id: string;
  name: string;
  slug: string;
  category: ProductCategory;
  description: string;
  use_case: string;
  price_min: number;
  price_max: number;
  moq: number;
  moq_unit: string;
  is_smartstock: boolean;
  is_eco: boolean;
  sample_tier: ProductSampleTier;
  sample_price: number;
  image_url?: string | null;
  specs?: ProductSpecs;
  compliance_certs?: string[];
  delivery_days_india: number;
  delivery_days_global: number;
  is_active: boolean;
  created_at?: string;
}